loremIpsum moved into Typography.mocks.tsx
